+++ src/auth/auth.routes.js
@@ -1,8 +1,10 @@
 const express = require("express");
 const { registerController, loginController } = require("./auth.controller");
+const { registerValidator, loginValidator } = require("./auth.validator");
+
 const router = express.Router();
 
-router.post("/register", registerController);
-router.post("/login", loginController);
+router.post("/register", registerValidator, registerController);
+router.post("/login", loginValidator, loginController);
 
 module.exports = router;
